client/requirements.txt
pillow
numpy
cryptography
brotli
colorlog
xxhash
Events
bcrypt
pillow-avif-plugin